refactor: consolidate token-accounting in sonar-sim; add output stream seam

See merge request web3-dex/GitHub-Sonar!81

+++ crates/sonar-sim/src/funding/common.rs
@@ -91,7 +91,7 @@ pub(super) fn update_token_amount_account<T: TokenAmountMut>(
         owner: *owner,
         decimals,
         amount_raw,
-        ui_amount: raw_to_ui_amount(amount_raw, decimals),
+        ui_amount: raw_to_ui_amount(amount_raw as i128, decimals),
         program_kind,
     })
 }
